fix: don't treat explicit + operations as string interpolations
Closes https://github.com/decaffeinate/decaffeinate/issues/425

There was already some logic to distinguish between explicit `+` operations and
implicit ones generated as part of string interpolation parsing, but that logic
wasn't used in all places. Now, we also use it in `isInterpolatedString`, which
avoids a problem where `+` was seen as a string interpolation in some cases.
