+++ test/examples/string-interpolation-plus-normal-string/input.coffee
@@ -1,0 +1,1 @@
+"#{a}" + "b"
