+++ test/examples/string-interpolation-plus-normal-string/output.json
@@ -1,0 +1,90 @@
+{
+  "type": "Program",
+  "line": 1,
+  "column": 1,
+  "range": [
+    0,
+    13
+  ],
+  "raw": "\"#{a}\" + \"b\"\n",
+  "body": {
+    "type": "Block",
+    "line": 1,
+    "column": 1,
+    "range": [
+      0,
+      12
+    ],
+    "statements": [
+      {
+        "type": "PlusOp",
+        "line": 1,
+        "column": 1,
+        "range": [
+          0,
+          12
+        ],
+        "left": {
+          "type": "TemplateLiteral",
+          "line": 1,
+          "column": 1,
+          "range": [
+            0,
+            6
+          ],
+          "raw": "\"#{a}\"",
+          "quasis": [
+            {
+              "type": "String",
+              "line": 1,
+              "column": 1,
+              "range": [
+                0,
+                1
+              ],
+              "data": "",
+              "raw": "\""
+            },
+            {
+              "type": "String",
+              "data": "",
+              "raw": "\"",
+              "line": 1,
+              "column": 6,
+              "range": [
+                5,
+                6
+              ]
+            }
+          ],
+          "expressions": [
+            {
+              "type": "Identifier",
+              "line": 1,
+              "column": 4,
+              "range": [
+                3,
+                4
+              ],
+              "data": "a",
+              "raw": "a"
+            }
+          ]
+        },
+        "right": {
+          "type": "String",
+          "line": 1,
+          "column": 10,
+          "range": [
+            9,
+            12
+          ],
+          "data": "b",
+          "raw": "\"b\""
+        },
+        "raw": "\"#{a}\" + \"b\""
+      }
+    ],
+    "raw": "\"#{a}\" + \"b\""
+  }
+}
